Correct misnamed rule tests
Fix #828.
