--- go/lang/correctness/permissions/openfile.go
@@ -1,23 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"os"
-)
-
-func main() {
-	// ruleid: incorrect-default-permission
-	_, err := os.OpenFile("/tmp/thing", os.O_CREATE|os.O_WRONLY, 0666)
-	if err != nil {
-		fmt.Println("Error opening a file!")
-		return
-	}
-}
-func main1() {
-	// ok
-	_, err := os.OpenFile("/tmp/thing", os.O_CREATE|os.O_WRONLY, 0600)
-	if err != nil {
-		fmt.Println("Error opening a file!")
-		return
-	}
-}
