--- go/lang/security/audit/crypto/use_of_weak_crypto_des.go
@@ -1,23 +1,0 @@
-package main
-
-import (
-	"crypto/des"
-)
-
-func main() {
-	// NewTripleDESCipher can also be used when EDE2 is required by
-	// duplicating the first 8 bytes of the 16-byte key.
-	ede2Key := []byte("example key 1234")
-
-	var tripleDESKey []byte
-	tripleDESKey = append(tripleDESKey, ede2Key[:16]...)
-	tripleDESKey = append(tripleDESKey, ede2Key[:8]...)
-	// ruleid: use-of-DES
-	_, err := des.NewTripleDESCipher(tripleDESKey)
-	if err != nil {
-		panic(err)
-	}
-
-	// See crypto/cipher for how to use a cipher.Block for encryption and
-	// decryption.
-}
