--- go/lang/security/audit/crypto/use_of_weak_crypto_md5.go
@@ -1,32 +1,0 @@
-package main
-
-import (
-	"crypto/md5"
-	"fmt"
-	"io"
-	"log"
-	"os"
-)
-
-func main() {
-	f, err := os.Open("file.txt")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer f.Close()
-
-	defer func() {
-		err := f.Close()
-		if err != nil {
-			log.Printf("error closing the file: %s", err)
-		}
-	}()
-
-	// ruleid: use-of-md5
-	h := md5.New()
-	if _, err := io.Copy(h, f); err != nil {
-		log.Fatal(err)
-	}
-	// ruleid: use-of-md5
-	fmt.Printf("%x", h.Sum(nil))
-}
