--- python/django/best-practice/django_os.py
@@ -1,5 +1,0 @@
-import django
-import os
-
-#ruleid: use-django-environ
-my_variable = os.environ["TESTING"]
