--- python/django/best-practice/os_django.py
@@ -1,5 +1,0 @@
-import os
-import django
-
-#ruleid: use-django-environ
-my_variable = os.environ["TESTING"]
