+++ python/django/best-practice/upsell_django_environ.yaml
@@ -2,6 +2,7 @@ rules:
 - id: use-django-environ
   patterns:
   - pattern-not: |
+      ...
       import environ
       ...
   - pattern-either:
